+++ src/main/php/auth/Roles.php
@@ -12,7 +12,6 @@ namespace stubbles\webapp\auth;
  * Represents a list of roles.
  *
  * @since  5.0.0
- * @ProvidedBy(stubbles\webapp\auth\ioc\RolesProvider.class)
  */
 class Roles implements \Countable, \IteratorAggregate
 {